Register injected identifier for originally unnamed default exports ('_DefaultExportValue') in Babel scope tracker
This also helps clear the following console warnings when used with plugin-transform-typescript:

The exported identifier "_DefaultExportValue" is not declared in Babel's scope tracker
as a JavaScript value binding, and "@babel/plugin-transform-typescript"
never encountered it as a TypeScript type declaration.
It will be treated as a JavaScript value.

This problem is likely caused by another plugin injecting
"_DefaultExportValue" without registering it in the scope tracker. If you are the author
 of that plugin, please use "scope.registerDeclaration(declarationPath)"